fix: 커스텀 삼진커널 dtype 수정 P003 완료 기록
[fix]
- ternary_kernel backward: autocast 밖에서 wq가 float32가 되어 bf16 @ float32 충돌 → dtype 통일
  (활성=bf16, 가중치 grad=float32). _ref_matmul 도 alpha를 활성 dtype로 캐스팅.
- trainer: label 정의를 로그 출력 이전으로 이동(UnboundLocalError 수정)

[feat]
- 분리형 커스텀 삼진 커널(ternary_kernel.py, 기본 off) + Triton forward(검증 후) + 안전 폴백
- 토큰-마크 명명 스냅샷(--snapshot-at), WSD decay-branch(--sched stable/decay, --decay-from)
- 후반부 EMA(--ema-start), KD 교사 선택(--kd-best), g 오버라이드(--mlp-group), weight centering(--center-weights)
- compare: 처방 문구 제거, '손실 격차 X (작음/보통/큼)' 만 표시

[docs/plan/result]
- test_result/002: P003 결과 — KD가 g=8 격차 소멸(+0.005, 2.09x 감축), 새 최적점 t_kd_g8
- test_plan P011~P014(저토큰교사·데이터효율·seq-warmup·커널검증), docs/custom_ternary_kernel

+++ tinylm/model/ternary_kernel.py
@@ -74,7 +74,7 @@ if _HAS_TRITON:
 def _ref_matmul(x, codes, alpha, group, dtype):
     """레퍼런스: dequant 후 F.linear. 기존 경로와 동일 결과(정확성 기준)."""
     O = codes.shape[0]
-    wq = (codes.to(dtype) * alpha).reshape(O, -1)      # [O, I]
+    wq = (codes.to(dtype) * alpha.to(dtype)).reshape(O, -1)   # [O, I] (활성 dtype로 통일)
     return F.linear(x, wq)
 
 
@@ -112,10 +112,11 @@ class _TernaryKernelLinear(torch.autograd.Function):
     def backward(ctx, gy):
         x, aw, alpha, codes = ctx.saved_tensors
         O, I, group, clip = ctx.meta
-        wq = (codes.to(gy.dtype) * alpha).reshape(O, I)         # [O, I]
-        gx = gy @ wq                                            # [.., I]
-        gw_full = gy.reshape(-1, O).t().to(x.dtype) @ x.reshape(-1, I)   # [O, I]
-        win = 1.0 / (1.0 + (aw / (clip * alpha).clamp_min(1e-8)).pow(4))  # [O, G, group]
+        wq = (codes.to(gy.dtype) * alpha.to(gy.dtype)).reshape(O, I)   # 활성 dtype로 통일
+        gx = gy @ wq                                            # [.., I] (bf16)
+        # 가중치 grad 는 float32(파라미터 dtype)로 정확히
+        gw_full = gy.reshape(-1, O).t().float() @ x.reshape(-1, I).float()
+        win = (1.0 / (1.0 + (aw / (clip * alpha).clamp_min(1e-8)).pow(4))).float()
         gw = (gw_full.reshape(O, I // group, group) * win).reshape(O, I)
         return gx, gw, None, None, None, None
